[refact] create new controllers for ip and mac endpoints
